pyproject.toml
[project]
name = "data-streaming"
version = "0.0.1"
description = "this is data streaming application"
authors = [
    {name = "Rupinder",email = "rupinderpal.singh@netsmartz.net"}
]
license = {text = "3"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "apache-airflow (>=2.10.5,<3.0.0)",
    "kafka-python (>=2.0.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "docker (>=7.1.0,<8.0.0)",
    "cassandra-driver (>=3.29.2,<4.0.0)",
    "spark (>=0.2.1,<0.3.0)",
    "pyspark (>=3.5.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
